+++ src/App.js
@@ -1,6 +1,6 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Switch, Route } from "react-router";
-import firebase, { firestore } from "./util/api/fbInstance";
+import { firestore } from "./util/api/fbInstance";
 import Header from "./components/Header/Header";
 import Footer from "./components/Footer/Footer";
 import {
@@ -15,6 +15,15 @@ import {
 } from "./components/pages";
 function App() {
   const [posts, setPosts] = useState([{ Title: "", Content: "" }]);
+  useEffect(() => {
+    firestore.collection("faq").onSnapshot((snapshot) => {
+      const postArray = snapshot.docs.map((doc) => ({
+        id: doc.id,
+        ...doc.data(),
+      }));
+      setPosts(postArray);
+    });
+  }, []);
   return (
     <div>
       <Header />
@@ -29,11 +38,7 @@ function App() {
         <Route exact path="/Faqeditpage" component={FaqEditPage} />
         <Route exact path="/introduce" component={Introduce} />
         <Route exact path="/product" component={Product} />
-        <Route
-          exact
-          path="/FaqPreviewPage"
-          render={() => <FaqPreviewPage posts={posts} />}
-        />
+        <Route exact path="/FaqPreviewPage" component={FaqPreviewPage} />
         <Route exact path="/detail" component={Detail} />
       </Switch>
       <Footer />
