+++ src/components/pages/FaqPreviewPage.jsx
@@ -1,9 +1,9 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import Inner from "../atoms/Inner";
 import styled from "styled-components";
-import { useHistory } from "react-router";
-import { Link } from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
 import queryString from "query-string";
+import { firestore } from "../../util/api/fbInstance";
 
 const FaqTitle = styled.p`
   margin-top: 50px;
@@ -45,6 +45,8 @@ const ContentBox = styled.div`
   height: 800px;
   border: 0.5px solid #000000;
   margin-left: 33px;
+  padding: 10px;
+  font-size: 1.25rem;
 `;
 const ListButton = styled.button`
   width: 7.5%;
@@ -58,25 +60,36 @@ const ListButtonWrap = styled.div`
   text-align: center;
 `;
 
-function FaqPreviewPage({ posts }) {
-  console.log(posts);
+function FaqPreviewPage() {
+  const [read, setRead] = useState([]);
   const history = useHistory();
   const { search } = history.location;
+
+  useEffect(() => {
+    firestore.collection("faq").onSnapshot((snapshot) => {
+      const postArray = snapshot.docs.map((doc) => ({
+        id: doc.id,
+        ...doc.data(),
+      }));
+      setRead(postArray);
+    });
+  }, []);
   const query = queryString.parse(search);
-  const faqState = posts.filter((item) => item.id === Number(query.id));
-  const { Title, Content } = faqState[0];
+  const faqState = read?.filter((item) => {
+    return item.id === query.id;
+  });
 
   return (
     <Inner>
       <FaqTitle>FAQ</FaqTitle>
       <TitleWrap>
         <TitleText>제목</TitleText>
-        <TitleContent>{Title?.map((item) => ({ item }))}</TitleContent>
+        <TitleContent>{faqState[0]?.Title}</TitleContent>
       </TitleWrap>
       <TitleWrap2>
         <TitleText2>내용</TitleText2>
       </TitleWrap2>
-      <ContentBox>{Content?.map((item) => ({ item }))}</ContentBox>
+      <ContentBox>{faqState[0]?.Content}</ContentBox>
       <ListButtonWrap>
         <Link to="/faq">
           <ListButton>목록</ListButton>
